More type fixes
git-svn-id: https://svn.code.sf.net/p/gaussianbeam/code@91 37a9f6fc-9735-0410-a863-98db706a8500

+++ gui/OpticsView.cpp
@@ -287,7 +287,7 @@ void OpticsView::setOrigin(double origin)
 	if (origin > sceneRect().right() - m_horizontalRange)
 		origin = sceneRect().right() - m_horizontalRange;
 
-	double value = origin*m_horizontalRuller->rullerScale();
+	int value = int(origin*m_horizontalRuller->rullerScale());
 	horizontalScrollBar()->setValue(value);
 	m_opticsViewProperties->setViewOrigin(origin);
 }
